refactor: исправила расположение виджетов
