+++ src/page-items/home/ui/HomePage.tsx
@@ -1,10 +1,9 @@
-import { AboutUs } from '@/page-items/home/ui/AboutUs';
-import { Events } from '@/page-items/home/ui/Events';
-
 import { GetMenu } from '@/features/menu';
 import { GetPopularPizza } from '@/features/popular';
-import { Preview } from '@/page-items/home/ui/Preview';
 
+import { AboutUs } from '@/widgets/about';
+import { Events } from '@/widgets/events';
+import { Preview } from '@/widgets/preview';
 
 
 export function HomePage() {
