+++ src/widgets/about/index.ts
@@ -1,0 +1,3 @@
+import { AboutUs } from './ui';
+export { AboutUs };
+
