+++ src/widgets/events/index.ts
@@ -1,0 +1,4 @@
+import { Events } from './ui';
+
+export { Events };
+
