+++ src/widgets/events/ui/EventItem/EventItem.props.ts
