+++ src/widgets/events/ui/EventItem/index.module.css
