+++ src/widgets/events/ui/EventItem/index.tsx
