+++ src/widgets/events/ui/Events.props.ts
