+++ src/widgets/events/ui/index.module.css
