+++ src/widgets/preview/index.ts
@@ -1,0 +1,3 @@
+import { Preview } from './ui';
+
+export { Preview };
