+++ src/widgets/preview/ui/CustomVideoPlayer/index.module.css
