+++ src/widgets/preview/ui/CustomVideoPlayer/index.tsx
