+++ src/widgets/preview/ui/CustomVideoPlayer/ui/PlayerItem.tsx
