+++ src/widgets/preview/ui/CustomVideoPlayer/ui/index.module.css
