+++ src/widgets/preview/ui/DoubledButton/DoubledButton.props.ts
