+++ src/widgets/preview/ui/index.module.css
